separated NullAwayTests.java
