+++ nullaway/src/test/java/com/uber/nullaway/NullAwayContractsTests.java
@@ -1,0 +1,290 @@
+package com.uber.nullaway;
+
+import com.google.errorprone.CompilationTestHelper;
+import java.util.Arrays;
+import java.util.List;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+public class NullAwayContractsTests {
+  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+  private CompilationTestHelper defaultCompilationHelper;
+
+  /**
+   * Creates a new {@link CompilationTestHelper} with a list of javac arguments. As of Error Prone
+   * 2.5.1, {@link CompilationTestHelper#setArgs(List)} can only be invoked once per object. So,
+   * this method must be used to create a test helper when a different set of javac arguments is
+   * required than those used for {@link #defaultCompilationHelper}.
+   *
+   * @param args the javac arguments
+   * @return the test helper
+   */
+  private CompilationTestHelper makeTestHelperWithArgs(List<String> args) {
+    return CompilationTestHelper.newInstance(NullAway.class, getClass()).setArgs(args);
+  }
+
+  @Test
+  public void checkContractPositiveCases() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber",
+                "-XepOpt:NullAway:CheckContracts=true"))
+        .addSourceFile("CheckContractPositiveCases.java")
+        .doTest();
+  }
+
+  @Test
+  public void checkContractNegativeCases() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber",
+                "-XepOpt:NullAway:CheckContracts=true"))
+        .addSourceFile("CheckContractNegativeCases.java")
+        .doTest();
+  }
+
+  @Test
+  public void basicContractAnnotation() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber"))
+        .addSourceLines(
+            "NullnessChecker.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import org.jetbrains.annotations.Contract;",
+            "public class NullnessChecker {",
+            "  @Contract(\"_, null -> true\")",
+            "  static boolean isNull(boolean flag, @Nullable Object o) { return o == null; }",
+            "  @Contract(\"null -> false\")",
+            "  static boolean isNonNull(@Nullable Object o) { return o != null; }",
+            "  @Contract(\"null -> fail\")",
+            "  static void assertNonNull(@Nullable Object o) { if (o != null) throw new Error(); }",
+            "}")
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "class Test {",
+            "  String test1(@Nullable Object o) {",
+            "    return NullnessChecker.isNonNull(o) ? o.toString() : \"null\";",
+            "  }",
+            "  String test2(@Nullable Object o) {",
+            "    return NullnessChecker.isNull(false, o) ? \"null\" : o.toString();",
+            "  }",
+            "  String test3(@Nullable Object o) {",
+            "    NullnessChecker.assertNonNull(o);",
+            "    return o.toString();",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void impliesNonNullContractAnnotation() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber"))
+        .addSourceLines(
+            "NullnessChecker.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import org.jetbrains.annotations.Contract;",
+            "public class NullnessChecker {",
+            "  @Contract(\"!null -> !null\")",
+            "  static @Nullable Object noOp(@Nullable Object o) { return o; }",
+            "}")
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "class Test {",
+            "  String test1(@Nullable Object o1) {",
+            "    // BUG: Diagnostic contains: dereferenced expression",
+            "    return NullnessChecker.noOp(o1).toString();",
+            "  }",
+            "  String test2(Object o2) {",
+            "    return NullnessChecker.noOp(o2).toString();",
+            "  }",
+            "  Object test3(@Nullable Object o1) {",
+            "    // BUG: Diagnostic contains: returning @Nullable expression",
+            "    return NullnessChecker.noOp(o1);",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void malformedContractAnnotations() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber"))
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import org.jetbrains.annotations.Contract;",
+            "class Test {",
+            "  @Contract(\"!null -> -> !null\")",
+            "  static @Nullable Object foo(@Nullable Object o) { return o; }",
+            "  @Contract(\"!null -> !null\")",
+            "  static @Nullable Object bar(@Nullable Object o, String s) { return o; }",
+            "  @Contract(\"jabberwocky -> !null\")",
+            "  static @Nullable Object baz(@Nullable Object o) { return o; }",
+            // We don't care as long as nobody calls the method:
+            "  @Contract(\"!null -> -> !null\")",
+            "  static @Nullable Object dontcare(@Nullable Object o) { return o; }",
+            "  static Object test1() {",
+            "    // BUG: Diagnostic contains: Invalid @Contract annotation",
+            "    return foo(null);",
+            "  }",
+            "  static Object test2() {",
+            "    // BUG: Diagnostic contains: Invalid @Contract annotation",
+            "    return bar(null, \"\");",
+            "  }",
+            "  static Object test3() {",
+            "    // BUG: Diagnostic contains: Invalid @Contract annotation",
+            "    return baz(null);",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void contractNonVarArg() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber"))
+        .addSourceLines(
+            "NullnessChecker.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import org.jetbrains.annotations.Contract;",
+            "public class NullnessChecker {",
+            "  @Contract(\"null -> fail\")",
+            "  static void assertNonNull(@Nullable Object o) { if (o != null) throw new Error(); }",
+            "}")
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "class Test {",
+            "  void test(java.util.function.Function<Object, Object> fun) {",
+            "    NullnessChecker.assertNonNull(fun.apply(new Object()));",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void contractPureOnlyIgnored() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber"))
+        .addSourceLines(
+            "PureLibrary.java",
+            "package com.example.library;",
+            "import org.jetbrains.annotations.Contract;",
+            "public class PureLibrary {",
+            "  @Contract(",
+            "    pure = true",
+            "  )",
+            "  public static String pi() {",
+            "    // Meh, close enough...",
+            "    return Double.toString(3.14);",
+            "  }",
+            "}")
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import com.example.library.PureLibrary;",
+            "import javax.annotation.Nullable;",
+            "class Test {",
+            "  String piValue() {",
+            "    String pi = PureLibrary.pi();",
+            "    // Note: we must trigger dataflow to ensure that",
+            "    // ContractHandler.onDataflowVisitMethodInvocation is called",
+            "    if (pi != null) {",
+            "       return pi;",
+            "    }",
+            "    return Integer.toString(3);",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void customContractAnnotation() {
+    makeTestHelperWithArgs(
+            Arrays.asList(
+                "-d",
+                temporaryFolder.getRoot().getAbsolutePath(),
+                "-XepOpt:NullAway:AnnotatedPackages=com.uber",
+                "-XepOpt:NullAway:CustomContractAnnotations=com.example.library.CustomContract",
+                "-XepOpt:NullAway:CheckContracts=true"))
+        .addSourceLines(
+            "CustomContract.java",
+            "package com.example.library;",
+            "import static java.lang.annotation.RetentionPolicy.CLASS;",
+            "import java.lang.annotation.Retention;",
+            "@Retention(CLASS)",
+            "public @interface CustomContract {",
+            "  String value();",
+            "}")
+        .addSourceLines(
+            "NullnessChecker.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import com.example.library.CustomContract;",
+            "public class NullnessChecker {",
+            "  @CustomContract(\"_, !null -> !null\")",
+            "  @Nullable",
+            "  static Object bad(Object a, @Nullable Object b) {",
+            "    if (a.hashCode() % 2 == 0) {",
+            "      // BUG: Diagnostic contains: Method bad has @Contract",
+            "      return null;",
+            "    }",
+            "    return new Object();",
+            "  }",
+            "",
+            "  @CustomContract(\"_, !null -> !null\")",
+            "  @Nullable",
+            "  static Object good(Object a, @Nullable Object b) {",
+            "    if (a.hashCode() % 2 == 0) {",
+            "      return b;",
+            "    }",
+            "    return new Object();",
+            "  }",
+            "}")
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "class Test {",
+            "  String test1() {",
+            "    return NullnessChecker.good(\"bar\", \"foo\").toString();",
+            "  }",
+            "  String test2() {",
+            "    // BUG: Diagnostic contains: dereferenced expression",
+            "    return NullnessChecker.good(\"bar\", null).toString();",
+            "  }",
+            "}")
+        .doTest();
+  }
+}
