+++ nullaway/src/test/java/com/uber/nullaway/NullAwayTryConstructsTests.java
@@ -1,0 +1,90 @@
+package com.uber.nullaway;
+
+import com.google.errorprone.CompilationTestHelper;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+public class NullAwayTryConstructsTests {
+  @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
+
+  private CompilationTestHelper defaultCompilationHelper;
+
+  @Test
+  public void tryFinallySupport() {
+    defaultCompilationHelper.addSourceFile("NullAwayTryFinallyCases.java").doTest();
+  }
+
+  @Test
+  public void tryWithResourcesSupport() {
+    defaultCompilationHelper
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import java.io.BufferedReader;",
+            "import java.io.FileReader;",
+            "import java.io.IOException;",
+            "class Test {",
+            "  String foo(String path, @Nullable String s, @Nullable Object o) throws IOException {",
+            "    try (BufferedReader br = new BufferedReader(new FileReader(path))) {",
+            "      // Code inside try-resource gets analyzed",
+            "      // BUG: Diagnostic contains: dereferenced expression",
+            "      o.toString();",
+            "      s = br.readLine();",
+            "      return s;",
+            "    }",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void tryWithResourcesSupportInit() {
+    defaultCompilationHelper
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import java.io.BufferedReader;",
+            "import java.io.FileReader;",
+            "import java.io.IOException;",
+            "class Test {",
+            "  private String path;",
+            "  private String f;",
+            "  Test(String p) throws IOException {",
+            "    path = p;",
+            "    try (BufferedReader br = new BufferedReader(new FileReader(path))) {",
+            "      f = br.readLine();",
+            "    }",
+            "  }",
+            "}")
+        .doTest();
+  }
+
+  @Test
+  public void tryFinallySupportInit() {
+    defaultCompilationHelper
+        .addSourceLines(
+            "Test.java",
+            "package com.uber;",
+            "import javax.annotation.Nullable;",
+            "import java.io.BufferedReader;",
+            "import java.io.FileReader;",
+            "import java.io.IOException;",
+            "class Test {",
+            "  private String path;",
+            "  private String f;",
+            "  Test(String p) throws IOException {",
+            "    path = p;",
+            "    try {",
+            "      BufferedReader br = new BufferedReader(new FileReader(path));",
+            "      f = br.readLine();",
+            "    } finally {",
+            "      f = \"DEFAULT\";",
+            "    }",
+            "  }",
+            "}")
+        .doTest();
+  }
+}
